Simulator/requirements.txt
Flask==0.12.1
Flask-Sockets==0.2.1
numpy==1.14.0
scipy==1.0.0
azure-storage==0.36.0